Refactor preference classes into new package.
