+++ app/src/main/java/tech/ula/model/repositories/AppsRepository.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.withContext
 import tech.ula.model.daos.AppsDao
 import tech.ula.model.entities.App
 import tech.ula.model.remote.RemoteAppsSource
-import tech.ula.utils.AppsPreferences
+import tech.ula.utils.preferences.AppsPreferences
 
 class AppsRepository(
     private val appsDao: AppsDao,
