+++ app/src/main/java/tech/ula/model/state/AppsStartupFsm.kt
@@ -11,6 +11,10 @@ import tech.ula.model.entities.Filesystem
 import tech.ula.model.entities.Session
 import tech.ula.model.repositories.UlaDatabase
 import tech.ula.utils.* // ktlint-disable no-wildcard-imports
+import tech.ula.utils.preferences.AppServiceTypePreference
+import tech.ula.utils.preferences.AppsPreferences
+import tech.ula.utils.preferences.PreferenceHasNotBeenSelected
+import tech.ula.utils.preferences.SshTypePreference
 
 class AppsStartupFsm(
     ulaDatabase: UlaDatabase,
