+++ app/src/main/java/tech/ula/utils/DownloadUtility.kt
@@ -6,6 +6,7 @@ import org.rauschig.jarchivelib.Archiver
 import org.rauschig.jarchivelib.ArchiverFactory
 import tech.ula.R
 import tech.ula.model.repositories.DownloadMetadata
+import tech.ula.utils.preferences.AssetPreferences
 import java.io.File
 import java.io.IOException
