+++ app/src/main/java/tech/ula/viewmodel/MainActivityViewModel.kt
@@ -14,6 +14,7 @@ import tech.ula.model.entities.Session
 import tech.ula.model.repositories.DownloadMetadata
 import tech.ula.model.state.* // ktlint-disable no-wildcard-imports
 import tech.ula.utils.* // ktlint-disable no-wildcard-imports
+import tech.ula.utils.preferences.AppServiceTypePreference
 import java.io.FileNotFoundException
 import kotlin.coroutines.CoroutineContext
