+++ app/src/test/java/tech/ula/model/state/AppsStartupFsmTest.kt
@@ -21,6 +21,10 @@ import tech.ula.model.entities.Filesystem
 import tech.ula.model.entities.Session
 import tech.ula.model.repositories.UlaDatabase
 import tech.ula.utils.* // ktlint-disable no-wildcard-imports
+import tech.ula.utils.preferences.AppsPreferences
+import tech.ula.utils.preferences.PreferenceHasNotBeenSelected
+import tech.ula.utils.preferences.SshTypePreference
+import tech.ula.utils.preferences.VncTypePreference
 import java.io.IOException
 
 @RunWith(MockitoJUnitRunner::class)
