+++ app/src/test/java/tech/ula/viewmodel/MainActivityViewModelTest.kt
@@ -19,6 +19,7 @@ import tech.ula.model.entities.Session
 import tech.ula.model.repositories.DownloadMetadata
 import tech.ula.model.state.* // ktlint-disable no-wildcard-imports
 import tech.ula.utils.* // ktlint-disable no-wildcard-imports
+import tech.ula.utils.preferences.SshTypePreference
 import java.io.FileNotFoundException
 import java.lang.IllegalStateException
