+++ build.gradle
@@ -7,8 +7,8 @@ buildscript {
     }
 
     ext {
-        kotlin_version = '1.3.4'
-        android_plugin = '3.4.1'
+        kotlin_version = '1.3.41'
+        android_plugin = '3.4.2'
         jacoco_version = '0.8.4'
         navigation_version = '2.1.0-alpha05'
     }
